Fix: FaceData holds radius properties in Double
The protocol uses 8 bytes to represent each value.
It's 'Double', not 'Float'
(It's confusing that python's float is actually 'double' in other language as it's 8 byte long.)

+++ FaceDataServer/Type/FaceData.cs
@@ -12,15 +12,15 @@ namespace Cjbc.FaceDataServer.Type {
     public class FaceData {
         /// <summary>face x rotation in radian.</summary>
         /// <remarks>This should be range of (-<c>Math.PI</c>, <c>Math.PI</c>)</remarks>
-        public readonly float FaceXRadian;
+        public readonly double FaceXRadian;
 
         /// <summary>face y rotation in radian.</summary>
         /// <remarks>This should be range of (-<c>Math.PI</c>, <c>Math.PI</c>)</remarks>
-        public readonly float FaceYRadian;
+        public readonly double FaceYRadian;
 
         /// <summary>face z rotation in radian.</summary>
         /// <remarks>This should be range of (-<c>Math.PI</c>, <c>Math.PI</c>)</remarks>
-        public readonly float FaceZRadian;
+        public readonly double FaceZRadian;
 
         /// <summary>how much percent current mouth height is comparing to the default.</summary>
         /// <remarks>This should be range of (0, 150)</remarks>
@@ -37,7 +37,7 @@ namespace Cjbc.FaceDataServer.Type {
         /// <remarks>This should be range of (0, 150)</remarks>
         public readonly byte   RightEyePercent;
 
-        public FaceData(float x, float y, float z, byte mh, byte mw, byte le, byte re) {
+        public FaceData(double x, double y, double z, byte mh, byte mw, byte le, byte re) {
             FaceXRadian = x;
             FaceYRadian = y;
             FaceZRadian = z;
